Add required files to create a custom cloud image
Fedora is using imagefactory to build the official cloud image
published upstream. The idea of this patch is to build a similar
image but using the stable repo to build it and setting as only
repository the stable one instead of officially released by
fedora.

kickstart is taken from https://pagure.io/fedora-kickstarts/

Change-Id: I3a151cc31a57cdcb8bb53870d86b7f62c8261883